refactor: move TxnStatus and TrackingIdResponse to tracking_id.go
- Moved the TxnStatus enum and its JSON marshaling/unmarshaling methods from types_paymaster.go to tracking_id.go for better organization.
